Initial commit of source files & basic components
Initial files from create-react-app added to the project. The initial
components for the calculator have been added, but still need their
functionality added and to be styled. Clicking on any of the number
components passes the value back up to the main App component which
will have methods created for processing the click. App.js maintains
control of all the states within the app.