Add functionality for +- key and refactor addNum
The +- key has been added and logic written. Is it created as a
NumKey component as the logic is more of a toggle state and would have
created more compelx code and exclusion if routed through the
operations method. The addNum method logic was rewritten to exclude
this key from the normal concatenation process and a separate
conditional created to process this particular key.

Additionally, the initial state of currVal was changed to null rather
than "0". This allowed for removing and mixing some logical operators
within the addNum method without affecting functionality.