Add logic for clearing display and refactor code
The button to clear the display and necessary logic was created.
Additional logic was written into the NumKey component to update the
value of the key depending on the value props passed down. Logic was
also introduced to keep components that are not the clear button from
re-rendering each time the state changes. Since only the clear button
relies on the display state changing to dynamically update from "C" to
"AC" there was no need to have the other components update.

The logic for the plus/minus key was also refactored to address a bug
when the state.currVal was 0.