Combine NumKey and MathKey components
The MathKey and NumKey components have been combined into one new Key
component. All of the functions and props are passed down to the
component, but only the clear button instance will update with the
changes to state.

The onClick function of each component is assigned based on the
declared type of the button in the App.js component, just above the
render call.

This resolved maintaing to CSS files for essentially the same component