Resolve issues with +-  after using equals key
A second conditional had to be created to allow this key to be used if
a user had pressed the equals key to complet an operation. This was due
to the data being stored in a different state after completing the
equals operation versus other operations. This also applies to use
cases where the percentage key is used as the initial operator.