Fix bug that would keep active operator highlighted
There was a bug wherein the user would use the clear button mid operation, and
then click the equals button. In this case, the operator would remain
highlighted and allow for multiple active operators. In order to resolve an
additional if statement was added to the equals switch statement which removes
the active class from the operator if it exists.