Fix #1 with overflow display for larger numbers
This resolves the bug for the display overflowing with longer numnbers.
In order to resolve this issue, the display is capped at 16 characters
in length for inputting new numbers. If the display is already at 16
characters it will not register new number keypresses.

When completing equations where the resultant number is longer than 16
the app will now adjust the display state to be under 16 characters
using scientific notation. It will also add to existing scientific
notation if present.

Lastly, a separate if statement catches if there is a "+" included in
the scientific notation that may be defaulted by the JavaScript math
equation.