Fix bug in logic after a user presses equals
There was a bug when a user would press equals, then clear the result,
and then try using an operand. This would throw an error because no
operand was stored in state. To resolve this, a separate conditional
was introduced to return the calculator to its initial state if the
clear button is clicked after a user presses the equals button.