Fix bug when number is pressed following "="
Some of the states would get misaligned and throw an error if a user
tried to start a new equation immediately after pressing equals due to
there already being a value stored in the prevVal state but no operator
stored in state. To resolve this bug, a reset of the states takes place
within this condition, thereby treating the states as if starting from
initial state.